Use datetime_display property in api response to figure out time quiz was generated.
Why?
We were previously using the quiz title. For some reason the api is sometimes including a comma in the title such as "Quiz: Morning trivia challenge: April, 27, 2022". Because the comma between april and 27 isn't consistent our parsing method only works some times.
This change uses the datetime_display property instead which comes with a date formatted like so: "05:00 27/04/2022"